+++ packages/block-brokers/src/trustless-gateway/trustless-gateway.ts
@@ -19,6 +19,7 @@ export interface TransformRequestInit {
 export interface TrustlessGatewayComponents {
   logger: ComponentLogger
   transformRequestInit?: TransformRequestInit
+  routing: string
 }
 
 export interface GetRawBlockOptions {
@@ -78,10 +79,13 @@ export class TrustlessGateway {
   private readonly log: Logger
   private readonly transformRequestInit?: TransformRequestInit
 
-  constructor (url: URL | string, { logger, transformRequestInit }: TrustlessGatewayComponents) {
+  public readonly routing: string
+
+  constructor (url: URL | string, { logger, transformRequestInit, routing }: TrustlessGatewayComponents) {
     this.url = url instanceof URL ? url : new URL(url)
     this.transformRequestInit = transformRequestInit
-    this.log = logger.forComponent(`helia:trustless-gateway-block-broker:${this.url.hostname}`)
+    this.log = logger.forComponent(`helia:trustless-gateway-block-broker:${this.url.host}`)
+    this.routing = routing
   }
 
   /**
@@ -151,7 +155,7 @@ HTTP/1.1 %d %s
 
           if (!res.ok) {
             this.#errors++
-            throw new Error(`Unable to fetch raw block for CID ${cid} from gateway ${this.url}, recieved ${res.status} ${res.statusText}`)
+            throw new Error(`Unable to fetch raw block for CID ${cid} from gateway ${this.url}, received ${res.status} ${res.statusText}`)
           }
           // limited Response ensures the body is less than 2MiB (or configurable maxSize)
           // see https://github.com/ipfs/helia/issues/790