+++ packages/utils/test/fixtures/create-helia.ts
@@ -1,3 +1,4 @@
+import { defaultLogger } from '@libp2p/logger'
 import { MemoryBlockstore } from 'blockstore-core'
 import { MemoryDatastore } from 'datastore-core'
 import { isLibp2p, createLibp2p } from 'libp2p'
@@ -17,7 +18,9 @@ export async function createHelia (opts: Partial<HeliaInit> = {}): Promise<Helia
   } else if (opts.libp2p != null) {
     libp2p = await createLibp2p(opts.libp2p)
   } else {
-    libp2p = stubInterface<Libp2p<any>>()
+    libp2p = stubInterface<Libp2p<any>>({
+      logger: defaultLogger()
+    })
   }
 
   const init = {