+++ packages/utils/test/pins.depth-limited.spec.ts
@@ -1,4 +1,3 @@
-/* eslint-env mocha */
 import { expect } from 'aegir/chai'
 import { MemoryBlockstore } from 'blockstore-core'
 import drain from 'it-drain'