+++ packages/unixfs/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## [7.0.4](https://github.com/ipfs/helia/compare/unixfs-v7.0.3...unixfs-v7.0.4) (2026-02-05)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^6.1.0 to ^6.1.1
+
+## [7.0.3](https://github.com/ipfs/helia/compare/unixfs-v7.0.2...unixfs-v7.0.3) (2026-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^6.0.3 to ^6.1.0
+
+## [7.0.2](https://github.com/ipfs/helia/compare/unixfs-v7.0.1...unixfs-v7.0.2) (2026-02-02)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^6.0.2 to ^6.0.3
+
 ## [7.0.1](https://github.com/ipfs/helia/compare/unixfs-v7.0.0...unixfs-v7.0.1) (2026-01-15)