+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,55 @@
 # Changelog
 
+## [2.4.2](https://github.com/ipfs/helia/compare/utils-v2.4.1...utils-v2.4.2) (2026-02-05)
+
+
+### Bug Fixes
+
+* allow configuring cid peer filter size ([#955](https://github.com/ipfs/helia/issues/955)) ([e16b7a7](https://github.com/ipfs/helia/commit/e16b7a7dcd013f13321dea162eee0130473541ea))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^6.1.0 to ^6.1.1
+
+## [2.4.1](https://github.com/ipfs/helia/compare/utils-v2.4.0...utils-v2.4.1) (2026-02-05)
+
+
+### Bug Fixes
+
+* emit provider event for initial providers ([#953](https://github.com/ipfs/helia/issues/953)) ([f058fba](https://github.com/ipfs/helia/commit/f058fba8a06f8afdd3558bdf7c3801381648ffec))
+* emit provider event for initial provides ([f058fba](https://github.com/ipfs/helia/commit/f058fba8a06f8afdd3558bdf7c3801381648ffec))
+
+## [2.4.0](https://github.com/ipfs/helia/compare/utils-v2.3.5...utils-v2.4.0) (2026-02-04)
+
+
+### Features
+
+* allow adding peers to session ([#950](https://github.com/ipfs/helia/issues/950)) ([33e4681](https://github.com/ipfs/helia/commit/33e4681394539d4298a028d2d9ff48a14f76a8e3))
+
+
+### Bug Fixes
+
+* add name to block brokers ([#949](https://github.com/ipfs/helia/issues/949)) ([0456c42](https://github.com/ipfs/helia/commit/0456c42dbd92d94633c133d4f5fe35264a6bbb80))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^6.0.3 to ^6.1.0
+
+## [2.3.5](https://github.com/ipfs/helia/compare/utils-v2.3.4...utils-v2.3.5) (2026-02-02)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^6.0.2 to ^6.0.3
+
 ## [2.3.4](https://github.com/ipfs/helia/compare/utils-v2.3.3...utils-v2.3.4) (2026-01-15)